Remove old page button. Reorg files.
